further minor changes to snmptranslate report
Back in commit 3a614fc we discovered that net-snmp will read its index files
and then process MIBs in the listed order, making the snmptranslate reports
sensitive to this ordering. However the indexes are written in the order MIBs
are found in a MIBDIR, which is entirely filesystem dependent when using
readdir(3), as net-snmp does (it would be better to use scandir(3), I guess).

Anyway, this commit is the final tidying up of the snmptranslate report after
the fix in 3a614fc. There are a small number of changes but they are all
caused by a vendor redefining a leaf within their own MIBs, and if we load
their MIBs in a different order, then... well we can't be blamed if the report
changes as well.