Make Aruba 3810 series switches naming consistent with other families
With the former ProCurve family of switches being put under the Aruba brand
name, model naming for newer switches changes from HP switch to Aruba switch.
The MIBs released by HP are not entirely consistent in object naming; the
3810 series was the first one to be released under the Aruba brand name.

Naming for the 3810 series is not consistent with the HP ProCurve branded
models, nor with the 2930 and 2540 series that were added shortly after under
the Aruba name. Change naming for 3810 switch models to be more consistent
with Aruba 2930 and 2540 series.